Change the frequency of timer

+++ device/timer.c
@@ -20,6 +20,6 @@ static void frequency_set(uint8_t counter_port, uint8_t counter_no, uint8_t rwl,
 
 void timer_init(){
   put_str("timer_init start\n");
-  frequency_set(COUNTER0_PORT, COUNTER0_NO, READ_WRITE_LATCH, COUNTER_MODE, COUNTER0_VALUE);
+  frequency_set(COUNTER0_PORT, COUNTER0_NO, READ_WRITE_LATCH, COUNTER_MODE, (uint8_t)COUNTER0_VALUE);
   put_str("timer_init done\n");
 }
